[JENKINS-43845] - Move classes to the util package
